% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank_initialize.R
\name{draw_lambda0}
\alias{draw_lambda0}
\title{Draw initial lambda}
\usage{
draw_lambda0(n_tenure, s0 = 0.1, delta = 0.1)
}
\arguments{
\item{n_tenure}{The length of a tenure of a driver.}

\item{s0}{Varianve of the initial condition. Mean of the initial condition is set to 0.}

\item{delta}{Variace of the innovation.}
}
\description{
Draw initial lambda
}
\keyword{internal}
