% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank_initialize.R
\name{draw_lambda_th}
\alias{draw_lambda_th}
\title{Draw initial lambda_type}
\usage{
draw_lambda_th(lambda_mean, sigma)
}
\arguments{
\item{lambda_mean}{A vector of prior means with length corresponding to the length of a tenure.}

\item{sigma}{Variace of the emission equation.}
}
\description{
Draw initial lambda_type
}
\keyword{internal}
