% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank-data.R
\name{dyRank_data}
\alias{dyRank_data}
\title{Data preparation for dynamic rating estimation}
\usage{
dyRank_data(data, var_rank, var_player, var_match, var_time,
  var_rank_type = NULL)
}
\arguments{
\item{data}{A data frame}
}
\description{
Data preparation for dynamic rating estimation
}
