% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank_initialize.R
\name{dyRank_initialize_params}
\alias{dyRank_initialize_params}
\title{parameter initialization for dyRank()}
\usage{
dyRank_initialize_params(data)
}
\description{
parameter initialization for dyRank()
}
\keyword{internal}
