% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank.R
\name{dyRank_single}
\alias{dyRank_single}
\title{Hierarchical Dynamic Rating (Single variance)}
\usage{
dyRank_single(dat_GP, time_GP, year_GP, delta, Y, player_fix, n_fs, tr,
  mcmc, burn, thin, verboser = TRUE)
}
\description{
Hierarchical Dynamic Rating (Single variance)
}
