% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank-dlm.R
\name{FFBS}
\alias{FFBS}
\title{Sample Global Mean Parameter via FFBS}
\usage{
FFBS(lambda_list, tau, m0, s0, delta)
}
\arguments{
\item{lambda_list}{A list of time-specific mean parameter.}

\item{tau}{A variance parameter of the emission model.}

\item{m0}{Mean parameter of the initial condition.}

\item{s0}{Variance parameter of the initial condition.}

\item{delta}{An innovation parameter (variance) for the state equation.}
}
\value{
A vector of dynamic mean value of length T (the length of the career)
}
\description{
Sample Global Mean Parameter via FFBS
}
\examples{
# generate data 
set.seed(1234)
n_tenure <- 20
lambda   <- rep(NA, n_tenure)
lambda_list <- vector('list', length = n_tenure)
tau       <- runif(1, 0.5, 1.5)
delta     <- 0.5
for (i in 1:n_tenure) {
  if (i == 1) {
    # initial condition 
    lambda[1] <- rnorm(1, 0, 0.5)
  } else {
    lambda[i] <- rnorm(1, mean = lambda[i-1], sd = sqrt(delta))    
  }

  lambda_list[[i]] <- rnorm(5, mean = lambda[i], sd = tau)
}

# fit 
lambda_fit <- FFBS(lambda_list, tau = tau, m0 = 0, s0 = 0.5^2, delta = delta)

# plot 
plot(lambda, type = 'o', pch = 16, ylim = c(-3, 4))
lines(lambda_fit, type = 'o', pch = 21, col = 'red')
lines(sapply(lambda_list, mean), type = 'o', pch = 17)
}
\keyword{internal}
