% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FFBScpp}
\alias{FFBScpp}
\title{C++ implementation of FFBB algorithm}
\usage{
FFBScpp(lambda_mat, tau, m0 = 0, s0 = 0.5, delta = 0.5)
}
\description{
C++ implementation of FFBB algorithm
}
\keyword{internal}
