% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FFBSsingle_cpp}
\alias{FFBSsingle_cpp}
\title{C++ implementation of FFBB algorithm}
\usage{
FFBSsingle_cpp(Z, Omega, is_fix, m0 = 0, s0 = 0.25, delta = 0.25)
}
\description{
C++ implementation of FFBB algorithm
}
\keyword{internal}
