% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank-summary.R
\name{get_mcmc}
\alias{get_mcmc}
\title{Obtain MCMC object of rating parameters}
\usage{
get_mcmc(obj)
}
\arguments{
\item{obj}{A \code{dyRank.fit} object, which is an output of \code{\link{dyRank}} or \code{\link{hdyRank}}.}
}
\value{
A list of mcmc objects that are compatible with functions from \code{coda} package.
}
\description{
Obtain MCMC object of rating parameters
}
