% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdyRank.R
\name{hdyRank}
\alias{hdyRank}
\title{Hierarchical Dynamic Rating Estimation for Ranked Data}
\usage{
hdyRank(
  data,
  var_rank,
  var_player,
  var_match,
  var_time,
  var_rank_type,
  driver_fix,
  mcmc = 100,
  burnin = 10,
  thin = 1,
  truncation = 3
)
}
\arguments{
\item{data}{A data frame of \code{tibble} class.}

\item{var_rank}{A variable name of the outcome (ranking).}

\item{var_player}{A variable name of the players.}

\item{var_match}{A variable name of matches.}

\item{var_time}{A variable name of time index.}

\item{var_rank_type}{A variable name of rank types.}

\item{driver_fix}{A name of the player used as a refenrece.}

\item{mcmc}{MCMC iterations.}

\item{burnin}{Burnin periods.}

\item{thin}{Thinning.}

\item{truncation}{A truncation parameter.}
}
\description{
\code{hdyRank()} implements the hierarhical dynamic Plackett-Luce model
 for estimating the rating based on rank-ordered data. 
 Different from \code{dyRank()}, \code{hdyRank()} can handle multiple types of ranking data.
}
