% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialize_counts_cpp}
\alias{initialize_counts_cpp}
\title{Initialize c_mk}
\usage{
initialize_counts_cpp(dat, lambda, race_attr, driver_attr, n_race)
}
\arguments{
\item{dat}{A matrix}

\item{lambda}{A matrix of new values of λ}

\item{race_attr}{A matrix of race attributes.}
}
\value{
Updated c_mk.
}
\description{
Initialize c_mk
}
