% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank_initialize.R
\name{initialize_params}
\alias{initialize_params}
\title{Function to initialize parameters}
\usage{
initialize_params(data)
}
\arguments{
\item{data}{An output of \code{dyRank_data()}.}
}
\description{
Function to initialize parameters
}
