% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank-summary.R
\name{plot_rating}
\alias{plot_rating}
\title{Plot rating estimates}
\usage{
plot_rating(obj, facet = FALSE, driver_name = NULL, y_label = "",
  x_label = "", ncol = 3)
}
\arguments{
\item{obj}{An output}
}
\description{
Plot rating estimates
}
