% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank-hd-sampler.R
\name{sample_lambda}
\alias{sample_lambda}
\title{Sample Ability Parameters}
\usage{
sample_lambda(dat_GP, time_GP, delta, Y, lambda_mean, lambda, tau, c_mk,
  n_tenure, driver_fix, tr, a0 = 1.5, b0 = 1)
}
\description{
Sample Ability Parameters
}
\keyword{internal}
