% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_counts_cpp}
\alias{update_counts_cpp}
\title{C++ implementation of count updates}
\usage{
update_counts_cpp(c_mk, dat_GP, time_GP, delta, lambda, tr)
}
\description{
C++ implementation of count updates
}
\keyword{internal}
