% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_cov_cpp}
\alias{update_cov_cpp}
\title{Update variance-covariance matrix}
\usage{
update_cov_cpp(lambda_mat, lambda_mean, v0, S0)
}
\arguments{
\item{lambda_mat}{Observed λ matrix.}

\item{v0}{Degree of freedom parameter of the inverse Wishart (prior).}

\item{S0}{Scale matrix of the inverse Wishart (prior).}
}
\value{
A single draw from the inverse Wishart distribution.
}
\description{
Update variance-covariance matrix
}
\keyword{internal}
