% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank-hd-sampler.R
\name{update_moments}
\alias{update_moments}
\title{Data augmentation}
\usage{
update_moments(dat_GP, time_GP, delta, Y, lambda, c_mk, nt, tr = 1)
}
\description{
Data augmentation
}
\keyword{internal}
