% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_moments_cpp}
\alias{update_moments_cpp}
\title{Update moments for a driver}
\usage{
update_moments_cpp(dat, race_attr, driver_attr, lambda, c_mk)
}
\arguments{
\item{dat}{A matrix of one driver's information.}
}
\description{
Update moments for a driver
}
\keyword{internal}
