% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank.R
\name{dyRank}
\alias{dyRank}
\title{Hierarchical Dynamic Rating}
\usage{
dyRank(data, var_rank, var_player, var_match, var_time, var_rank_type,
  driver_fix, mcmc = 100, burnin = 10, thin = 1, n_chains = 3,
  compute_gelman = FALSE)
}
\description{
Hierarchical Dynamic Rating
}
