% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank.R
\name{dyRank}
\alias{dyRank}
\title{Hierarchical Dynamic Rating}
\usage{
dyRank(dat_GP, time_GP, year_GP, delta, Y, player_fix, n_fs, tr, mcmc,
  burn, thin, verboser = TRUE)
}
\description{
Hierarchical Dynamic Rating
}
