% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyRank.R
\name{dyRank}
\alias{dyRank}
\title{Dynamic Ranking Model}
\usage{
dyRank(
  data,
  var_rank,
  var_player,
  var_match,
  var_time,
  driver_fix,
  mcmc = 100,
  burnin = 10,
  thin = 1,
  truncation = 3
)
}
\arguments{
\item{data}{A data frame of \code{tibble} class.}

\item{var_rank}{A variable name of the outcome (ranking).}

\item{var_player}{A variable name of the players.}

\item{var_match}{A variable name of matches.}

\item{var_time}{A variable name of time index.}

\item{driver_fix}{A name of the player used as a refenrece.}

\item{mcmc}{MCMC iterations.}

\item{burnin}{Burnin periods.}

\item{thin}{Thinning.}

\item{truncation}{A truncation parameter.}
}
\value{
\code{dyRank()} returns an object of \code{dyRank.fit} class, which is a list
of the following elements:
\describe{
   \item{\code{lambda}}{A list of estimated rating parameters.
         Rating estimates can be accessed with \code{get_rating()} or
         \code{get_mcmc()} function.}
   \item{\code{data}}{A transformed dataset used for estimation.}
}
}
\description{
Dynamic Ranking Model
}
\examples{
set.seed(1234)
# load data
data("f1_race", package = "dyRank")

## estimate via MCMC
fit <- dyRank(
    data       = f1_race,
    var_rank   = "Pos",
    var_player = "driver",
    var_match  = "GP",
    var_time   = "year",
    driver_fix = "Timo Glock",
    mcmc = 10, burnin = 10, thin = 1,
    truncation = 3
)

## obtain estimated ratings
rating <- get_rating(fit)
}
\seealso{
\code{\link{get_rating}}, \code{\link{get_mcmc}}
}
