% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FFBSmult_cpp}
\alias{FFBSmult_cpp}
\title{FFBS for Multivariate Model}
\usage{
FFBSmult_cpp(lambda_mat, Sigma, is_fix, m0 = 0, s0 = 0.25, delta = 0.25)
}
\description{
FFBS for Multivariate Model
}
\keyword{internal}
