% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_moments}
\alias{update_moments}
\title{Update moments for a driver}
\usage{
update_moments(dat, race_attr, driver_attr, lambda, c_mk, trunc = 5L)
}
\arguments{
\item{dat}{A matrix of one driver's information.}
}
\description{
Update moments for a driver
}
\keyword{internal}
